Leaving the effects framework

+++ Samples/03 SceneTest/SceneMaterials/SceneMaterials.cs
@@ -365,7 +365,7 @@ namespace SceneTest.SceneMaterials
 
             for (int i = 0; i < count; i++)
             {
-                model[i].MaterialIndex = (uint)i;
+                model[i].MaterialIndex = i;
             }
 
             return model;
